Решение задачи 43

+++ MyLibs/MyMetods.cs
@@ -1,10 +1,10 @@
 ﻿namespace MyLibs;
 public class MyMetods
 {
-   public static int[] GetArrayFromString(string? arrayStr)
+   public static double[] GetArrayFromString(string? arrayStr)
    {
       string[] mass = arrayStr!.Split(new char[] { ' ', ',', '*' }, StringSplitOptions.RemoveEmptyEntries);
-      int[] result = new int[mass.Length];
+      double[] result = new double[mass.Length];
       for (int i = 0; i < mass.Length; i++)
       {
          result[i] = Convert.ToInt32(mass[i]);
